Moved object and multi index definition to its own header file and added type comment
